moved reply to tweet workflow and constants
