+++ echos_lab/twitter/twitter_client.py
@@ -7,7 +7,6 @@ import requests
 from sqlalchemy.orm import Session
 from tweepy import Response, Tweet, User
 
-from echos_lab.common.env import get_env
 from echos_lab.common.logger import logger
 from echos_lab.common.utils import with_db
 from echos_lab.db import db_connector
@@ -23,27 +22,11 @@ from echos_lab.twitter.types import (
     ReferenceTypes,
     TweetExclusions,
     TweetMention,
+    TWEET_FIELDS,
+    RESPONSE_RATING_THRESHOLD_FOLLOWERS,
+    RESPONSE_RATING_THRESHOLD_MENTIONS,
+    MEME_RATING_THRESHOLD,
 )
-
-# Threshold to quote tweet instead of reply tweet from reply guy
-# e.g. 0.2 means 20% of the time, the reply will be a quote tweet
-REPLY_GUY_QUOTE_TWEET_THRESHOLD = 0.05
-
-RESPONSE_RATING_THRESHOLD_MENTIONS = int(get_env("RESPONSE_RATING_THRESHOLD_MENTIONS", 4))
-RESPONSE_RATING_THRESHOLD_FOLLOWERS = int(get_env("RESPONSE_RATING_THRESHOLD_FOLLOWERS", 7))
-MEME_RATING_THRESHOLD = int(get_env("MEME_RATING_THRESHOLD", 9))
-
-# Percent of time that a reply to a followed account should be skipped (to prevent spam)
-# e.g. 0.2 means 20% of tweets are skipped
-FOLLOWER_SKIP_REPLY_THRESHOLD = 0.5
-
-# Fields to include when fetching tweets:
-#  - author_id: user ID of the author of the tweet
-#  - created_at: datetime of when the tweet was created
-#  - conversation_id: if the tweet is a reply, this will be the root tweet
-#      otherwise, this will be the same as the tweet ID
-#  - referenced_tweets: includes refernce tweet during quote tweets or replies
-TWEET_FIELDS = ["author_id", "created_at", "conversation_id", "referenced_tweets"]
 
 
 async def get_user_id_from_username(username: str) -> int | None:
@@ -836,53 +819,3 @@ async def reply_to_followers(db: Session, agent_profile: AgentProfile, tweets: l
             reply_to_tweet_id=tweet.tweet.tweet_id,
             quote_tweet_id=tweet.tweet.tweet_id,
         )
-
-
-@with_db
-async def reply_to_tweet(db: Session, agent_profile: AgentProfile, tweet_id: int) -> int | None:
-    """
-    Generates and posts a reply guy response to a specific tweet
-    Returns the tweet ID of the response
-    """
-    # Fetch the tweet
-    # TODO: use pipeline get_tweet_from_tweet_id
-    tweet = await get_tweet_from_tweet_id(tweet_id)
-    if not tweet:
-        return None
-
-    # Enrich the tweet with all the parents in the thread
-    # It's not techically a mention, but the struct has the needed functionality
-    mention = await enrich_user_mention(tweet)
-
-    # If this was a thread, grab the author of the top parent,
-    # otherwise grab the author last tweet in the reply
-    original_tweet = mention.tagged_tweet if not mention.original_tweet else cast(HydratedTweet, mention.original_tweet)
-    author = await original_tweet.get_username()
-
-    # Get the recent tweets from this author
-    author_recent_tweets = await get_author_recent_tweets(original_tweet.author_id)
-
-    # Generate a response
-    conversation_summary = await prompts.build_twitter_mentions_prompt(mention)
-    response_evaluation = await post_maker.generate_reply_guy_tweet(
-        agent_profile=agent_profile,
-        author=author,
-        tweet_summary=conversation_summary,
-        author_recent_tweets=author_recent_tweets,
-        agent_recent_tweets=[],
-        allow_roasting=True,
-    )
-
-    if response_evaluation is None:
-        return None
-
-    # Post the response
-    return await post_tweet_response(
-        agent_profile=agent_profile,
-        evaluation=response_evaluation,
-        meme_threshold=MEME_RATING_THRESHOLD,
-        text_threshold=RESPONSE_RATING_THRESHOLD_MENTIONS,
-        conversation_id=mention.tagged_tweet.conversation_id,
-        reply_to_tweet_id=mention.tagged_tweet.id,
-        quote_tweet_id=original_tweet.id,
-    )
